requirements.txt
# Core ML
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# RL
gymnasium>=0.28.0

# Data
yfinance>=0.2.0
ta>=0.10.0              # Technical analysis indicators

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0

# Optional: Dashboard
streamlit>=1.22.0

# Development
jupyter>=1.0.0
black>=23.0.0